refactor:PaymentRepository BigInteger -> Long

+++ src/main/java/com/example/shoppingmallproject/payment/repository/PaymentRepository.java
@@ -4,8 +4,7 @@ import com.example.shoppingmallproject.payment.entity.Payment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
 
 @Repository
-public interface PaymentRepository extends JpaRepository<Payment, BigInteger> {
+public interface PaymentRepository extends JpaRepository<Payment, Long> {
 }
